[C++] Warnings cleanup.

+++ aeron-client/src/main/c/concurrent/aeron_broadcast_transmitter.c
@@ -82,7 +82,7 @@ int aeron_broadcast_transmitter_transmit(
         insert_padding_record(
             (aeron_broadcast_record_descriptor_t *)(transmitter->buffer + record_offset), (int32_t)to_end_of_buffer);
 
-        current_tail += to_end_of_buffer;
+        current_tail += (int64_t)to_end_of_buffer;
         record_offset = 0;
     }
     else
